requirements.txt
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
flasgger==0.9.5
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
MarkupSafe==1.1.1
mistune==0.8.4
numpy==1.20.1
pandas==1.2.3
Pillow==8.1.2
PyJWT==1.7.1
pymongo==3.11.3
pyrsistent==0.17.3
pysqlite3==0.4.5
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
schedule==1.0.0
six==1.15.0
twilio==6.52.0
urllib3==1.26.3
Werkzeug==1.0.1
